Made RetryableDataSourceBeanPostProcessor a standalone class.
